requirements.txt
flask>=3.1.0
python-dotenv>=1.0.0
ibm-watson==6.1.0
ibm-watsonx-ai>=1.2.0
PyPDF2>=3.0.1
python-docx>=0.8.11
